fix: BadgeController, BadgeService, UserFriendController에서 Sequence로 변경/리턴하는 함수에서 NotFoundException예외처리가 안 되어 있는 서비스 수정, 컨트롤러의 불필요한 null check수정/Empty인경우 NO_CONTENT로 반환하도록한것 수정/예외처리 Exception Handler로 던지도록 수정

+++ a107/src/main/java/com/ssafy/a107/api/service/BadgeServiceImpl.java
@@ -80,8 +80,10 @@ public class BadgeServiceImpl implements BadgeService {
     }
 
     @Override
-    public void deleteBadge(Long seq) {
-        badgeRepository.deleteById(seq);
+    public void deleteBadge(Long seq) throws NotFoundException{
+        if(badgeRepository.existsById(seq)){
+            badgeRepository.deleteById(seq);
+        }else throw new NotFoundException("Invalid badge sequence!");
     }
